Moving PropertyGridView DropDownHolder control to a separate file.
